Window blinds work. Window count >2 works.
I've tested up to four windows, now.
All should be draw-able using their given surfaces, all of which are
deleted when the game is closed.

I'm not sure if I'd rather an invisible/visible close system or a true
create/delete one. For gMirage, though, it's best to go with
invisible/visible, since that's how VB6 handles it.