Remove debug code, test earlier changes.
The startup debug code is removed, since the issue it was added for has
been fixed.

The test windows have been modified to make use of the simpler window
code.

The game window was changed down to 800 by 600 to fit most screens
without a problem.

Interpolation was turned off. Hopefully this fixes some problems with
the window display.

Lastly, the theming was changed for no particular reason.